+++ CHANGELOG.md
@@ -1332,3 +1332,4 @@
 - [2026-05-22 18:30] fix(i18n): resolve race condition during i18n initialization
 - [2026-05-22 18:39] fix(navigation): resolve race condition during navigation initialization
 - [2026-05-22 18:41] test(perf): implement widget integration tests for perf
+- [2026-05-22 19:10] fix(theme): prevent double tap duplicate action in theme